作业

+++ ClassLibraryEF/Migrations/MyDbContextModelSnapshot.cs
@@ -21,9 +21,9 @@ namespace ClassLibraryEF.Migrations
 
             modelBuilder.Entity("Rbac.Entity.Admin", b =>
                 {
-                    b.Property<long>("AdminId")
+                    b.Property<int>("AdminId")
                         .ValueGeneratedOnAdd()
-                        .HasColumnType("bigint")
+                        .HasColumnType("int")
                         .HasAnnotation("SqlServer:ValueGenerationStrategy", SqlServerValueGenerationStrategy.IdentityColumn);
 
                     b.Property<int>("CreateId")
@@ -69,8 +69,8 @@ namespace ClassLibraryEF.Migrations
                         .HasColumnType("int")
                         .HasAnnotation("SqlServer:ValueGenerationStrategy", SqlServerValueGenerationStrategy.IdentityColumn);
 
-                    b.Property<long>("AdminId")
-                        .HasColumnType("bigint");
+                    b.Property<int>("AdminId")
+                        .HasColumnType("int");
 
                     b.Property<int>("RoleId")
                         .HasColumnType("int");
